refactor(daemon): split server/mod.rs into focused submodules
mod.rs grew to 2,084 LOC as new daemon features piled on. Split it
into topic-focused submodules so every file stays under 1,000 LOC and
the IPC server's surface area is easier to navigate.

mod.rs is now a 124-LOC glue file that defines DaemonServer, declares
every submodule, and re-exports their items via `use ...::*;` so each
sibling can keep its existing `use super::*;` pattern.

Mechanical move only — no behavior changes, no API changes. The 226
daemon lib tests + 10 ignored continue to pass.

New submodules under crates/zccache-daemon/src/server/:

- state.rs              SharedState (the daemon's central state object)
- lifecycle.rs          DaemonServer::{bind, bind_with_cache_dir}, accessors, test seams
- run.rs                DaemonServer::run main loop + watcher pipeline initializer
- cached_artifact.rs    CachedArtifact / CachedPayload + payload materialization + legacy .meta migration
- compiler_hash.rs      Compiler-binary hash memoization keyed by (mtime, size)
- request_cache.rs      RequestCacheEntry / CachedRequestPath / etc.
- rsp_cache.rs          Response-file (@file) expansion + caching
- cache_trim.rs         Time-based + size-capped trimmers for ephemeral caches
- in_flight.rs          RAII guard for state.in_flight_bytes
- pch.rs                PCH source-header resolution
- client_env.rs         Replay client env into compiler children
- session.rs            Request::SessionStart handler + session log writer
- handle_clear.rs       Request::Clear handler
- handle_compile_ephemeral.rs   Single-roundtrip ephemeral compile + run_compiler_direct
- watch.rs              watch_directory / watch_directories
- util.rs               persist_workers_default, hash_file, context_files_fresh, lookup_artifact_with_disk_fallback
